API 리셋 후 다시 만들기

+++ app/src/main/java/dev/kichan/marketplace/model/data/login/LoginRes.kt
@@ -1,5 +1,5 @@
 package dev.kichan.marketplace.model.data.login
 
 data class LoginRes(
-    val studentId : Int,
+    val studentId : Long,
 )
